infra-automation-cost-tool/cost-reporter/requirements.txt
azure-identity
azure-mgmt-costmanagement
